implemented authentication system in backend for clients and fixed authentication of admin panel and updated product model and renamed user file to authentication

+++ api/views/products.py
@@ -7,6 +7,7 @@ from models.product import Product
 from api.views import app_views
 from flask import abort, jsonify, make_response, redirect, request, current_app, url_for, flash
 from werkzeug.utils import secure_filename
+from api.utils.auth_utils import token_required
 from sys import stderr
 import sys, os
 
@@ -20,15 +21,41 @@ def get_products():
     list_products = []
     for product in all_products:
         dct = product.to_dict()
+        for key in dct.keys():
+            if key in ['created_at', 'updated_at', 'id']:
+                del dct[key]
+
+        list_products.append(dct)
+    return jsonify(list_products)
+
+@app_views.route('/products/full', methods=['GET'], strict_slashes=False)
+@token_required
+def get_products_full(current_user):
+    """
+    Retrieves the list of all product objects with full details for admins only
+    """
+    # check if user is admin or not
+    if not current_user.is_admin:
+        abort(401, description='Not allowed')
+
+    all_products = storage.all(Product).values()
+    list_products = []
+    for product in all_products:
+        dct = product.to_dict()
         list_products.append(dct)
     return jsonify(list_products)
 
 
 @app_views.route('/products/<id>', methods=['GET'], strict_slashes=False)
-def get_product_with_id(id=None):
+@token_required
+def get_product_with_id(current_user, id=None):
     """
     Retrieves the product with the id
     """
+    # check if user is admin or not
+    if not current_user.is_admin:
+        abort(401, description='Not allowed')
+
     # check if id is valid
     if id is None or id == '' or len(id) <= 0 or type(id) is not str:
         return make_response(jsonify({'error': 'the passed id is not of valid type'}), 400)
@@ -46,10 +73,15 @@ def get_product_with_id(id=None):
 
 
 @app_views.route('/products/<id>', methods=['PUT'], strict_slashes=False)
-def update_product_with_id(id=None):
+@token_required
+def update_product_with_id(current_user, id=None):
     """
     Update the product with the id
     """
+    # check if user is admin or not
+    if not current_user.is_admin:
+        abort(401, description='Not allowed')
+
     #get request body
     body = request.get_json()
 
@@ -91,10 +123,15 @@ def allowed_image(image):
 
 
 @app_views.route('/products', methods=['POST'], strict_slashes=False)
-def add_product():
+@token_required
+def add_product(current_user):
     """
     create a new product and store it in database
     """
+    # check if user is admin or not
+    if not current_user.is_admin:
+        abort(401, description='Not allowed')
+
     data = dict(request.form)
     if not data or type(data) is not dict:
         abort(400, description='Not a JSON')
@@ -135,10 +172,15 @@ def add_product():
     return redirect(url_for('app_views.admin_get_products'))
 
 @app_views.route('/products/delete/<id>', methods=['POST'], strict_slashes=False)
-def delete_product(id):
+@token_required
+def delete_product(current_user, id):
     """
     Retrieves the list of all product objects
     """
+    # check if user is admin or not
+    if not current_user.is_admin:
+        abort(401, description='Not allowed')
+
     product = storage.get(Product, id)
     if product:
         storage.delete(product)
